# backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.createEventRequest:
    properties:
      EndAt:
        description: RFC3339
        type: string
      EventOwner:
        type: string
      Location:
        type: string
      StartAt:
        description: RFC3339
        type: string
      Status:
        type: string
      Title:
        type: string
    type: object
  handlers.createProductRequest:
    properties:
      Category:
        type: string
      ProductName:
        type: string
      ProductQty:
        type: integer
      StockID:
        type: string
      Unit:
        type: string
    type: object
  handlers.createStockRequest:
    properties:
      StockName:
        type: string
      UserID:
        type: string
    type: object
  handlers.loginRequest:
    properties:
      Email:
        type: string
      Password:
        type: string
    type: object
  handlers.registerRequest:
    properties:
      AvatarURL:
        type: string
      DisplayName:
        type: string
      Email:
        type: string
      Password:
        type: string
    type: object
  models.Events:
    properties:
      CreatedAt:
        type: string
      EndAt:
        type: string
      EventID:
        type: string
      EventOwner:
        type: string
      EventOwnerName:
        type: string
      Location:
        type: string
      StartAt:
        type: string
      Status:
        type: string
      Title:
        type: string
    type: object
  models.Products:
    properties:
      Category:
        type: string
      ProductID:
        type: string
      ProductName:
        type: string
      ProductQty:
        type: integer
      StockID:
        type: string
      Unit:
        type: string
    type: object
  models.Stocks:
    properties:
      StockID:
        type: string
      StockName:
        type: string
      UserID:
        type: string
    type: object
  models.Users:
    properties:
      AvatarURL:
        type: string
      DisplayName:
        type: string
      Email:
        type: string
      Status:
        type: string
      UserId:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Public HTTP endpoints for the Event Blog backend.
  title: Event Blog API
  version: "1.0"
paths:
  /api:
    get:
      description: Lists available API endpoints and brief descriptions.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: API index
      tags:
      - meta
  /api/events:
    get:
      description: Returns all events.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Events'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Creates an event with owner, title, time range, and optional location/status.
      parameters:
      - description: Event data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.createEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Events'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new event
      tags:
      - events
  /api/health:
    get:
      description: Returns the current status of the API.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - meta
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticate user by email and password.
      parameters:
      - description: Login credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Users'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - users
  /api/products:
    get:
      description: Returns all products.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Products'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List products
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Creates a new product record.
      parameters:
      - description: Product data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.createProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Products'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a product
      tags:
      - products
  /api/register:
    post:
      consumes:
      - application/json
      description: Creates a new user document with a hashed password.
      parameters:
      - description: User registration data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Users'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - users
  /api/stocks:
    get:
      description: Returns all stocks.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Stocks'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List stocks
      tags:
      - stocks
    post:
      consumes:
      - application/json
      description: Creates a new stock record.
      parameters:
      - description: Stock data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.createStockRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Stocks'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a stock
      tags:
      - stocks
  /api/stocks/{stockId}:
    delete:
      description: Deletes a stock by ID and removes related products with the same
        StockID.
      parameters:
      - description: Stock ID (UUID)
        in: path
        name: stockId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a stock
      tags:
      - stocks
swagger: "2.0"